requirements.txt
beautifulsoup4==4.12.2
Flask==2.3.3
nltk==3.8.1
numpy==1.19.5
pandas==1.1.5
python_dateutil==2.8.2
requests==2.27.1
scikit_learn==1.3.0
spacy==3.6.1
text2emotion==0.0.5
textblob==0.17.1
wordcloud==1.9.2
